feat(web/ui): introduce reusable OpenChatComponent
Extracts the core AI chat UI logic into a new, highly customizable React component.
This component provides a drop-in solution for AI chat functionality with
OpenRouter and MCP tools, enhancing modularity and reusability.

The `index.tsx` route has been refactored to utilize this new component.
The `README.md` has been updated with detailed installation and usage examples,
including various customization options and prop documentation. New props like
`requireAuth`, `allowedModels`, and `renderMessage` have been added to enhance
flexibility.
