+++ apps/web/src/routes/__root.tsx
@@ -50,7 +50,7 @@ function RootComponent() {
 					disableTransitionOnChange
 					storageKey="vite-ui-theme"
 				>
-					<div className="grid grid-rows-[auto_1fr] h-svh grid-cols-1">
+					<div className="grid grid-rows-[auto_1fr] h-dvh grid-cols-1">
 						{isFetching ? <Loader /> : <Outlet />}
 					</div>
 					<Toaster richColors />
